% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19env.R
\docType{data}
\name{covid19_spain_2}
\alias{covid19_spain_2}
\title{COVID-19 Data for Spain.}
\format{
A dataframe with 1500 rows and 34 variables:
\describe{
  \item{province}{Names of provinces in Spain as factor}
  \item{CCAA}{Names of Autonomous Communities in Spain as factor}
  \item{ID_INE}{National Institute of Statistics unique identifier of the provinces}
  \item{Date}{Date}
  \item{Cases}{Number of COVID-19 cases reported in the province by date}
  \item{Incidence}{Incidence of COVID-19 cases per 100,000 population in the province by date}
  \item{Max_Temp}{Maximum temperature registered by the meteorological station representative of the province in Celsius}
  \item{Min_Temp}{Minimum temperature registered by the meteorological station representative of the province in Celsius}
  \item{Mean_Temp}{Mean temperature registered by the meteorological station representative of the province in Celsius}
  \item{Sunshine_Hours}{Hours of sunshine registered by the meteorological station representative of the province in hours}
  \item{Precipitation}{Precipitation registered by the meteorological station representative of the province}
  \item{Humidity}{Calculated humidity}
  \item{Mean_Temp_lag8}{8-day average of the mean temperature, calculated using the values between date-minus-12-days and date-minus-5-days in Celsius}
  \item{Sunshine_Hours_lag8}{8-day average of sunshine hours, calculated using the values between date-minus-12-days and date-minus-5-days in hours}
  \item{Precipitation_lag8}{8-day average of the precipitation, calculated using the values between date-minus-12-days and date-minus-5-days}
  \item{Humidity_lag8}{8-day average of the humidity, calculated using the values between date-minus-12-days and date-minus-5-days}
  \item{Mean_Temp_lag11}{11-day average of the mean temperature, calculated using the values between date-minus-12-days and date-minus-2-days in Celsius}
  \item{Sunshine_Hours_lag11}{11-day average of sunshine hours, calculated using the values between date-minus-12-days and date-minus-2-days in hours}
  \item{Precipitation_lag11}{11-day average of the precipitation, calculated using the values between date-minus-12-days and date-minus-2-days}
  \item{Humidity_lag11}{11-day average of the humidity, calculated using the values between date-minus-12-days and date-minus-2-days}
  \item{Mean_Temp_lag11w}{Wegihted 11-day average of the mean temperature, calculated using the values between date-minus-12-days and date-minus-2-days in Celsius}
  \item{Sunshine_Hours_lag11w}{Weighted 11-day average of sunshine hours, calculated using the values between date-minus-12-days and date-minus-2-days in hours}
  \item{Precipitation_lag11w}{Weighted 11-day average of the precipitation, calculated using the values between date-minus-12-days and date-minus-2-days}
  \item{Humidity_lag11w}{Weighted 11-day average of the humidity, calculated using the values between date-minus-12-days and date-minus-2-days}
  \item{geometry}{geometry of the simple features object}
}
}
\source{
Climatic data: Agencia Estatal de Meteorologia http://www.aemet.es/

COVID-19 data:

  Ministerio de Sanidad, Consumo y Bienestar Social: https://www.mscbs.gob.es/profesionales/saludPublica/ccayes/alertasActual/nCov-China/situacionActual.htm y gabinete de prensa del Ministerio de Sanidad.

  Departamento de Seguridad Nacional: https://www.dsn.gob.es/gl/current-affairs/press-room

  Instituto de Salud Carlos III. Situación de COVID-19 en España https://covid19.isciii.es/

  BOE: Crisis Sanitaria COVID-19 https://www.boe.es/biblioteca_juridica/codigos/codigo.php?id=355&modo=2&nota=0&tab=2

  Ministerio de Transportes, Movilidad y Agenda Urbana https://www.mitma.gob.es/transporte-terrestre/punto-de-informacion-de-servicios-de-restauracion

  Data collected by datadista https://github.com/datadista/datasets/tree/master/COVID%2019
}
\usage{
data(covid19_spain_2)
}
\description{
A dataset containing COVID-19 information for Spain, including
environmental correlates and control variables. The environmental variables
are calculated using an automated procedure that takes the average of all
meteorological stations inside a province.
}
\examples{
 data(covid19_spain_2)
 summary(covid19_spain_2)
}
\keyword{datasets}
