% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19env.R
\docType{package}
\name{covid19env}
\alias{covid19env}
\title{covid19env: A package with datasets for analysis of environmental correlates of COVID-19.}
\description{
This package includes datasets used for the analysis
of the climatic correlates of COVID-19 with their
respective documentation.
}
\references{
\url{https://github.com/paezha/Reproducible-Research-Workflow}
}
\author{
Fernando A. Lopez, Antonio Paez, School of Geography and Earth Sciences, McMaster University \email{paezha@mcmaster.ca}
}
