% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19env.R
\docType{data}
\name{provinces_spain}
\alias{provinces_spain}
\title{Provinces in Spain.}
\format{
A simple features dataframe with 52 rows and 4 columns:
\describe{
  \item{province}{Names of provinces in Spain as factor}
  \item{CCAA}{Names of Autonomous Communities in Spain as factor}
  \item{ID_INE}{National Institute of Statistics unique identifier of the provinces}
  \item{geometry}{geometry of the simple features object}
}
}
\source{
Instituto Nacional de Estadistica http://www.ine.es/
}
\usage{
data(provinces_spain)
}
\description{
A simple features object with the provinces in Spain.
}
\examples{
 data(provinces_spain)
 summary(provinces_spain)
}
\keyword{datasets}
